refactor(dispatch): extract envelope_adapters_provider.py leaf module (PR-6 monolith split)
Moves _fail_loud_on_empty_success, _map_generic_spawn_result, and
ProviderAdapter (incl. _run_kimi) out of dispatch_envelope.py unchanged.
ExecutionPlan and _AdapterResult are runtime imports (not TYPE_CHECKING) so
typing.get_type_hints() on ProviderAdapter.run/_run_kimi resolves against
THIS module's globals — the OI-288 failure mode, verified red then reverted.
The three function-local provider_dispatch imports stay function-local to
avoid reintroducing the cycle they were extracted to avoid.

dispatch_envelope.py re-exports ProviderAdapter at bare name so the
attribute-form coupling in test_dispatch_cli.py:756 keeps binding — verified
via an identity check (dispatch_envelope.ProviderAdapter is
envelope_adapters_provider.ProviderAdapter). Facade is now 511 lines (under
the 540-line DoD threshold) and contains only the lane register, LaneRouter,
the three entry points, and the re-export facade — this is the last PR of
fase 1.

Census expectation file regenerated: family gains envelope_adapters_provider,
layer4_classes/layer4_functions move their owner label for the four symbols;
couplings list itself is unchanged (test files import from dispatch_envelope,
which still resolves through the facade re-export).

Dispatch-ID: 20260804-005444-pr6-envelope-adapters-provider
