+++ tests/data/dispatch_envelope_census.json
@@ -816,6 +816,7 @@
   "family": [
     "dispatch_envelope",
     "envelope_adapters_claude",
+    "envelope_adapters_provider",
     "envelope_govern",
     "envelope_govern_support",
     "envelope_prepare",
@@ -823,9 +824,9 @@
   ],
   "layer4_classes": [
     "dispatch_envelope::LaneRouter",
-    "dispatch_envelope::ProviderAdapter",
     "envelope_adapters_claude::ClaudeSubprocessAdapter",
     "envelope_adapters_claude::CodexAdapter",
+    "envelope_adapters_provider::ProviderAdapter",
     "envelope_types::EnvelopeGovernError",
     "envelope_types::EnvelopeResult",
     "envelope_types::EnvelopeSpec",
@@ -833,15 +834,15 @@
   ],
   "layer4_functions": [
     "dispatch_envelope.LaneRouter::get",
-    "dispatch_envelope.ProviderAdapter::_run_kimi",
-    "dispatch_envelope.ProviderAdapter::run",
-    "dispatch_envelope::_fail_loud_on_empty_success",
-    "dispatch_envelope::_map_generic_spawn_result",
     "dispatch_envelope::run_envelope",
     "dispatch_envelope::run_envelope_headless_plan",
     "dispatch_envelope::run_envelope_plan",
     "envelope_adapters_claude.ClaudeSubprocessAdapter::run",
     "envelope_adapters_claude.CodexAdapter::run",
+    "envelope_adapters_provider.ProviderAdapter::_run_kimi",
+    "envelope_adapters_provider.ProviderAdapter::run",
+    "envelope_adapters_provider::_fail_loud_on_empty_success",
+    "envelope_adapters_provider::_map_generic_spawn_result",
     "envelope_govern::_govern",
     "envelope_govern_support::_archive_dispatch_events",
     "envelope_govern_support::_clear_dispatch_events",
